+++ src/store/epics/vendor.js
@@ -3,6 +3,7 @@ import { Observable } from 'rxjs';
 
 import types from '../constants';
 import actions from '../actions';
+import store from '../index';
 
 const vendor = {
 	vendorSave: (action$) =>
@@ -25,10 +26,7 @@ const vendor = {
 							return Observable.of(actions.onLoader(true), actions.vendorSaveSuccess(resp.response));
 						return Observable.of(actions.vendorSaveFailure('Something went wrong'));
 					})
-					.catch((err) => {
-						if (err.response) return Observable.of(actions.vendorSaveFailure(err.response));
-						return Observable.of(actions.vendorSaveFailure('Network Error'));
-					});
+					.catch(() => Observable.of(actions.vendorSaveFailure('Network Error')));
 			return Observable.of(actions.vendorSaveFailure('All fields are required'));
 		}),
 	getVendor: (action$) =>
@@ -50,13 +48,16 @@ const vendor = {
 						return Observable.of(actions.getVendorSuccess([]));
 					}
 				})
-				.catch((err) => {
-					if (err.response) return Observable.of(actions.getVendorFailure(err.response));
-					return Observable.of(actions.getVendorFailure('Network Error'));
-				});
+				.catch(() => Observable.of(actions.getVendorFailure('Network Error')));
 		}),
 	updateVendor: (action$) =>
 		action$.ofType(types.UPDATEVENDOR).switchMap(({ payload }) => {
+			const { purchases } = store.getState();
+			const vendorRecordFound = purchases.some((x) => x.vendorId._id === payload._id);
+			if (vendorRecordFound)
+				return Observable.of(
+					actions.updateVendorFailure("Transaction found for this vendor you can't edit it")
+				);
 			const isMatch = Object.values(payload).entries((val) => Boolean(val));
 			if (isMatch)
 				return Observable.ajax({
@@ -75,13 +76,17 @@ const vendor = {
 							return Observable.of(actions.onLoader(true), actions.updateVendorSuccess(resp.response));
 						return Observable.of(actions.updateVendorFailure('Something went Wrong'));
 					})
-					.catch(() => {
-						return Observable.of(actions.updateVendorFailure('Network Failure'));
-					});
+					.catch(() => Observable.of(actions.updateVendorFailure('Network Error')));
 			return Observable.of(actions.updateVendorFailure('All fields are required'));
 		}),
 	deleteVendor: (action$) =>
 		action$.ofType(types.DELETEVENDOR).switchMap(({ payload }) => {
+			const { purchases } = store.getState();
+			const vendorRecordFound = purchases.some((x) => x.vendorId._id === payload);
+			if (vendorRecordFound)
+				return Observable.of(
+					actions.updateVendorFailure("Transaction found for this vendor you can't delete it")
+				);
 			return Observable.ajax({
 				url: `http://localhost:8080/vendor/${payload}`,
 				method: 'DELETE',
@@ -96,9 +101,7 @@ const vendor = {
 					if (resp.response) return Observable.of(actions.deleteVendorSuccess(resp.response));
 					return Observable.of(actions.deleteVendorFailure('Something went Wrong'));
 				})
-				.catch(() => {
-					return Observable.of(actions.deleteVendorFailure('Network Failure'));
-				});
+				.catch(() => Observable.of(actions.deleteVendorFailure('Network Error')));
 		})
 };
